move code to exchange
